+++ src/Plugin.php
@@ -6,7 +6,6 @@ use benf\neo\controllers\Configurator as ConfiguratorController;
 use benf\neo\controllers\Conversion as ConversionController;
 use benf\neo\controllers\Input as InputController;
 use benf\neo\elements\Block;
-use benf\neo\enums\NewBlockMenuStyle;
 use benf\neo\fieldlayoutelements\ChildBlocksUiElement;
 use benf\neo\gql\interfaces\elements\Block as NeoGqlInterface;
 use benf\neo\integrations\feedme\Field as FeedMeField;
@@ -22,11 +21,11 @@ use craft\base\Plugin as BasePlugin;
 use craft\console\Application as ConsoleApplication;
 use craft\console\Controller;
 use craft\console\controllers\ResaveController;
-use craft\controllers\ElementsController;
 use craft\db\Query;
 use craft\db\Table;
+use craft\elements\conditions\SlugConditionRule;
+use craft\elements\GlobalSet;
 use craft\events\DefineConsoleActionsEvent;
-use craft\events\DefineElementEditorHtmlEvent;
 use craft\events\DefineFieldLayoutElementsEvent;
 use craft\events\RebuildConfigEvent;
 use craft\events\RegisterComponentTypesEvent;
@@ -40,7 +39,6 @@ use craft\gatsbyhelper\events\RegisterIgnoredTypesEvent;
 use craft\gatsbyhelper\services\Deltas;
 use craft\helpers\Console;
 use craft\helpers\Db;
-use craft\helpers\Html;
 use craft\models\FieldLayout;
 use craft\services\Fields;
 use craft\services\Gc;
@@ -124,7 +122,6 @@ class Plugin extends BasePlugin
         $this->_registerGatsbyHelper();
         $this->_registerFeedMeSupport();
         $this->_registerConditionFieldRuleRemoval();
-        $this->_registerDefaultBlockTypeIcon();
     }
 
     /**
@@ -245,7 +242,7 @@ class Plugin extends BasePlugin
                             ->select(['structureId'])
                             ->from('{{%neoblockstructures}}')
                             ->column(),
-                    ]
+                    ],
                 ],
             ]);
             $stdout("done\n", Console::FG_GREEN);
@@ -288,6 +285,11 @@ class Plugin extends BasePlugin
 
     private function _registerPermissions(): void
     {
+        // Only if the settings allow it
+        if (!$this->getSettings()->enableBlockTypeUserPermissions) {
+            return;
+        }
+
         Event::on(
             UserPermissions::class,
             UserPermissions::EVENT_REGISTER_PERMISSIONS,
@@ -356,29 +358,21 @@ class Plugin extends BasePlugin
                 if (self::$isGeneratingConditionHtml) {
                     $event->conditionRuleTypes = array_filter(
                         $event->conditionRuleTypes,
-                        fn($type) => !isset($type['fieldUid'])
-                    );
-                }
-            }
-        );
-    }
+                        function($type) use ($event) {
+                            // No field value conditions allowed as it may make existing blocks invalid
+                            if (isset($type['fieldUid'])) {
+                                return false;
+                            }
 
-    private function _registerDefaultBlockTypeIcon()
-    {
-        Event::on(
-            ElementsController::class,
-            ElementsController::EVENT_DEFINE_EDITOR_CONTENT,
-            function(DefineElementEditorHtmlEvent $event) {
-                if ($this->getSettings()->newBlockMenuStyle !== NewBlockMenuStyle::Classic && !$event->static) {
-                    $svg = Html::tag(
-                        'div',
-                        Html::modifyTagAttributes(
-                            Html::svg('@benf/neo/resources/default-new-block-icon.svg'),
-                            ['id' => 'ni-icon'],
-                        ),
-                        ['class' => 'hidden'],
+                            // Global sets don't have slugs
+                            if ($event->sender->elementType === GlobalSet::class && $type === SlugConditionRule::class) {
+                                return false;
+                            }
+
+                            // Everything else is okay
+                            return true;
+                        }
                     );
-                    $event->html = $svg . $event->html;
                 }
             }
         );