+++ src/assets/InputAsset.php
@@ -145,6 +145,8 @@ class InputAsset extends FieldAsset
         self::$filteredBlockTypes = $event->blockTypes;
 
         $jsSettings = [
+            'id' => $field->id,
+            'ownerId' => $owner?->id,
             'name' => $name,
             'namespace' => $view->namespaceInputName($name) . '[blocks]',
             'blockTypes' => self::_getBlockTypesJsSettings($field, $event->blockTypes, $owner),
@@ -175,6 +177,7 @@ class InputAsset extends FieldAsset
     private static function _getBlockTypesJsSettings(Field $field, array $blockTypes, ?ElementInterface $owner = null): array
     {
         $user = Craft::$app->getUser()->getIdentity();
+        $disablePermissions = !Neo::$plugin->getSettings()->enableBlockTypeUserPermissions;
         $jsBlockTypes = [];
 
         foreach ($blockTypes as $blockType) {
@@ -187,6 +190,7 @@ class InputAsset extends FieldAsset
                 $block->siteId = $owner->siteId;
             }
 
+            $ignorePermissions = $disablePermissions || $blockType->ignorePermissions;
             $jsBlockTypes[] = [
                 'id' => $blockType->id,
                 'sortOrder' => $blockType->sortOrder,
@@ -207,9 +211,9 @@ class InputAsset extends FieldAsset
                 'fieldLayoutId' => $blockType->fieldLayoutId,
                 'groupId' => $blockType->groupId,
                 'hasChildBlocksUiElement' => $blockType->hasChildBlocksUiElement(),
-                'creatableByUser' => $blockType->ignorePermissions || $user->can("neo-createBlocks:{$blockType->uid}"),
-                'deletableByUser' => $blockType->ignorePermissions || $user->can("neo-deleteBlocks:{$blockType->uid}"),
-                'editableByUser' => $blockType->ignorePermissions || $user->can("neo-editBlocks:{$blockType->uid}"),
+                'creatableByUser' => $ignorePermissions || $user->can("neo-createBlocks:{$blockType->uid}"),
+                'deletableByUser' => $ignorePermissions || $user->can("neo-deleteBlocks:{$blockType->uid}"),
+                'editableByUser' => $ignorePermissions || $user->can("neo-editBlocks:{$blockType->uid}"),
             ];
         }